ioctl cmd arg is u_long
SCCS-vsn: sys/netinet/in_var.h 8.2
SCCS-vsn: sys/netinet/in.c 8.4
SCCS-vsn: sys/netinet/tcp_usrreq.c 8.3
SCCS-vsn: sys/netinet/ip_var.h 8.2
SCCS-vsn: sys/netinet/udp_usrreq.c 8.5

+++ usr/src/sys/netinet/tcp_usrreq.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)tcp_usrreq.c	8.2 (Berkeley) %G%
+ *	@(#)tcp_usrreq.c	8.3 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -58,7 +58,7 @@ tcp_usrreq(so, req, m, nam, control)
 
 #if BSD>=43
 	if (req == PRU_CONTROL)
-		return (in_control(so, (int)m, (caddr_t)nam,
+		return (in_control(so, (u_long)m, (caddr_t)nam,
 			(struct ifnet *)control));
 	if (control && control->m_len) {
 		m_freem(control);
