+++ usr/src/sys/netinet/udp_usrreq.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)udp_usrreq.c	8.4 (Berkeley) %G%
+ *	@(#)udp_usrreq.c	8.5 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -451,7 +451,7 @@ udp_usrreq(so, req, m, addr, control)
 	int s;
 
 	if (req == PRU_CONTROL)
-		return (in_control(so, (int)m, (caddr_t)addr,
+		return (in_control(so, (u_long)m, (caddr_t)addr,
 			(struct ifnet *)control));
 	if (inp == NULL && req != PRU_ATTACH) {
 		error = EINVAL;
